% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP.R
\name{CP}
\alias{CP}
\title{Calculate conditional power}
\usage{
CP(
  r,
  n1 = NULL,
  n2 = NULL,
  alpha_1s,
  eff_est,
  eff_planned,
  eff_null,
  SE = NULL,
  type,
  pow = NULL,
  f = NULL
)
}
\arguments{
\item{r}{Proportion of subjects assigned to active group}

\item{n1}{Number of patients / events at first interim analysis (NULL if parameter f provided)}

\item{n2}{Number of patients / events at final analysis (NULL if parameter f provided)}

\item{alpha_1s}{one-sided alpha}

\item{eff_est}{effect observed/estimated at interim}

\item{eff_planned}{planned effect (usually alternative hypothesis, but can be modified to obtain \cr
conditional power for any assumprion)}

\item{eff_null}{effect corresponding with null hypothesis (e.g. 1 for hazard rates, 0 for difference) \cr
This accomodates for non-inferiority analyses}

\item{SE}{standard error: has to be provided if type not equal to "HR". If for instance SE =1 \cr
then eff_est and eff_planned correspond with z-scores}

\item{type}{if type="HR", then SE is calculated, if type="general", then SE's have to be provided \cr
by user}

\item{pow}{Power, needed to calculate sample size in second part provided to obtain given power}

\item{f}{n1/n2 at interim analysis. Only to be provided if n1 and n2 not provided}
}
\value{
a list of three vectors
\itemize{
\item CP_obs : conditional power, given 1) observed effect at interim, 2) total sample size, \cr
3) Assumed true effect= "eff_est" parameter
#' \item CP_planned : conditional power, given 1) observed effect at interim, 2) total sample size, \cr
3) Assumed true effect= "eff_planned" parameter
#' \item n2_inc_new : additional patients needed to obtain given power
}
}
\description{
This function calculates the conditional power at an interim analysis of a \cr
clinical trial
}
\examples{
CP(r=0.5,n1=72 ,n2=180,alpha_1s=0.025,eff_est=0.7,eff_planned=0.7 ,eff_null=1,SE=NULL,
   type="HR",pow=0.8)
CP(r=0.5              ,alpha_1s=0.025,eff_est=1.5,eff_planned=1.96,eff_null=0,SE=1,
   type="general",pow=0.8,f=0.5)
CP(r=0.5,n1=316,n2=633,alpha_1s=0.025,eff_est=1.5,eff_planned=1.96,eff_null=0,SE=1,
   type="general",pow=0.8)
}
\references{
Lan and Wittes. The B-Value: A Tool for Monitoring Data. Biometrics 1988;44:579-585 \cr
Mehta and Pocock
}
