% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP_PZ.R
\name{CP_PZ}
\alias{CP_PZ}
\title{Calculate promising zone and position of observed effect in promising zone}
\usage{
CP_PZ(
  r,
  n1 = NULL,
  n2 = NULL,
  alpha_1s,
  eff_est,
  eff_planned,
  eff_null,
  SE = NULL,
  p_c = NULL,
  sd_t = NULL,
  sd_c = NULL,
  type,
  pow = NULL,
  f = NULL,
  max,
  plot_effect
)
}
\arguments{
\item{r}{Proportion of subjects assigned to active group}

\item{n1}{Number of patients / events at first interim analysis (NULL if parameter f provided)}

\item{n2}{Number of patients / events at final analysis}

\item{alpha_1s}{one-sided alpha}

\item{eff_est}{effect observed/estimated at interim}

\item{eff_planned}{planned effect (usually alternative hypothesis, but can be modified to obtain \cr
conditional power for any assumprion)}

\item{eff_null}{effect corresponding with null hypothesis (e.g. 1 for hazard rates, 0 for difference) \cr
This accomodates for non-inferiority analyses}

\item{SE}{standard error: has to be provided if type not equal to "HR". If for instance SE =1 \cr
then eff_est and eff_planned correspond with z-scores}

\item{p_c}{proportion in control group (needs to be provided if type="prop")}

\item{sd_t}{standard deviation in treatment group (needs to be provided if type="cont")}

\item{sd_c}{standard deviation in control group (needs to be provided if type="cont")}

\item{type}{if type="HR", then SE is calculated, if type="general", then SE's have to be provided \cr
by user, if type="cont" (continuous) then sd_t and sd_c have to be provided \cr
if type="prop" then p_c has to be provided}

\item{pow}{Power, needed to calculate sample size in second part provided to obtain given power}

\item{f}{n1/n2 at interim analysis. Only to be provided if n1 and n2 not provided}

\item{max}{(Maximum sample size)/(original sample size), could be for instance 1.5, 2 or 3}

\item{plot_effect}{TRUE if plotting boundaries with corresponding effect scale}
}
\value{
a list of three vectors
\itemize{
\item CP_obs : conditional power, given 1) observed effect at interim, 2) total sample size, \cr
3) Assumed true effect= "eff_est" parameter
\item CP_planned : conditional power, given 1) observed effect at interim, 2) total sample size, \cr
3) Assumed true effect= "eff_planned" parameter
\item n2_inc_new : additional patients needed to obtain given power
\item CP_ll lower boundary of promising zone (z-score, b-value, conditional power, effect scale)
\item CP_ul upper boundary of promising zone (z-score, b-value, conditional power, effect scale)
}
}
\description{
This function calculates the promising zone boundaries at an interim analysis of a \cr
clinical trial
}
\examples{
CP_PZ(r=0.5,n1=72,n2=180,alpha_1s=0.025,eff_est=0.7,eff_planned=0.7,eff_null=1,SE=NULL,
      type="HR",max=1.5,pow=0.8       ,plot_effect=1)
CP_PZ(r=0.5      ,n2=180,alpha_1s=0.025,eff_est=0.7,eff_planned=0.7,eff_null=1,SE=NULL,
      type="HR",max=1.5,pow=0.8,f=0.25,plot_effect=1)
}
\references{
Lan and Wittes. The B-Value: A Tool for Monitoring Data. Biometrics 1988;44:579-585 \cr
Mehta CR, Pocock SJ. Adaptive increase in sample size when interim results are promising: A practical \cr
guide with examples. Statist. Med. 2011;30:3267–3284
}
